Clean up, improve camera screens
This commit is actually extremely large but alas.

Apart from improving the general data flow and streamlining the code a
bit, this forms the basis for the data flow and some other things.

Let's see how this goes. Right now the "Order" button after the QR
scanning makes the app crash because I don't think I have the segues
connected up the right way, but I'm looking into that.